# Improve text content argument handling