--- a/lib/phlex/component.rb
+++ b/lib/phlex/component.rb
@@ -5,32 +5,18 @@ module Phlex
     include Context
 
     module Overrides
-      module InitWithContent
-        def initialize(content = nil, *args, **kwargs, &block)
-          raise ArgumentError if block_given? && content
-
-          if block_given? && !block.binding.receiver.is_a?(Block)
-            block = Block.new(self, &block)
-          elsif content
-            block = Block.new(self) { text content }
-          end
-
-          @_content = block
-
-          super(*args, **kwargs)
+      def initialize(*args, **kwargs, &block)
+        if block_given? && !block.binding.receiver.is_a?(Block)
+          block = Block.new(self, &block)
         end
-      end
 
-      module InitWithoutContent
-        def initialize(*args, **kwargs, &block)
-          if block_given? && !block.binding.receiver.is_a?(Block)
-            block = Block.new(self, &block)
-          end
-
-          @_content = block
-
-          super(*args, **kwargs)
+        if !block_given? && args[0] && !method(__method__).super_method.parameters.to_h[:req]
+          content = args.delete_at(0)
+          block = Block.new(self) { text content }
         end
+
+        @_content = block
+        super(*args, **kwargs)
       end
 
       def template(...)
@@ -67,14 +53,7 @@ module Phlex
       end
 
       def inherited(child)
-        if child.instance_method(:initialize).parameters.to_h[:req]
-          child.prepend(Overrides::InitWithoutContent)
-        else
-          child.prepend(Overrides::InitWithContent)
-        end
-
         child.prepend(Overrides)
-
         super
       end
     end