Check if your initialiser handles positional arguments before assuming the first positional argument should be coerced into a block of text content.