@eregon 2022-08-11
IMHO this is too much magic (too hard to understand the behavior as it breaks encapsulation pretty drastically) and of course it's quite slow.
Also saw the `block.binding.receiver` above and that's also pretty slow and too magic IMHO.

If it's for fun, please go ahead, but if it's meant to be used project I would advise against getting any caller information and rather make things more explicit and less magic.