@joeldrapper 2022-08-11
> Also saw the `block.binding.receiver` above and that's also pretty slow and too magic IMHO.

The project lives or dies on this, as far as I can tell. But Phlex is already about 3x faster than ActionView and 32% faster than ViewComponent.