@eregon 2022-08-11
Did you measure the effect of this PR on perf? Would be interesting.

Re Proc#binding there are ideas floating around to remove it, because it causes unnecessary overhead for Ruby VMs.
But maybe it's not too bad perf-wise in this case.
From the code it seems you are using it to access the caller's self, right?