@joeldrapper 2022-08-11
> From the code it seems you are using it to access the caller's self, right?

Yes, that's right. We need a to-procable object that lets us access the caller.