@joeldrapper 2022-08-11
This PR seems to have no effect on our internal benchmark but it does negatively affect the the view-layer-benchmarks mentioned above.

```before
               phlex:     7047.8 i/s
      view_component:     5326.9 i/s - 1.32x  (± 0.00) slower
               cells:     3444.0 i/s - 2.05x  (± 0.00) slower
            partials:     2458.9 i/s - 2.87x  (± 0.00) slower
            dry_view:      345.9 i/s - 20.37x  (± 0.00) slower
```

After
```
               phlex:     6136.2 i/s
      view_component:     5379.7 i/s - 1.14x  (± 0.00) slower
               cells:     3494.1 i/s - 1.76x  (± 0.00) slower
            partials:     2468.5 i/s - 2.49x  (± 0.00) slower
            dry_view:      343.4 i/s - 17.87x  (± 0.00) slower
```