@joeldrapper 2022-08-11
If checking whether `initialize` takes positional arguments is slow, perhaps we could just memoize that on the class.

> IMHO this is too much magic (too hard to understand the behavior as it breaks encapsulation pretty drastically) and of course it's quite slow.

Let me explain the intended behaviour here.

If I define a component that takes block contents

```ruby
class CardComponent
  def template(&)
    div class: "rounded drop-shadow p-5", &
  end
end
```

I should be able to pass that component either a block:
```ruby
component(CardComponent) { h1 "Hello world!" }
```

Or a string:
```ruby
component CardComponent, "Hello world!"
```

But if I define a component that explicitly takes a positional argument

```ruby
class NavItem
  def initialize(label, to:)
    @label = label
    @to = to
  end

  def template
    li { a @label, to: @to }
  end
end
```

Then passing a positional argument shouldn't be received as a block but rather delegated directly to the original initialiser.

```ruby
component NavItem, "Home", "/"
```