@eregon 2022-08-12
Isn't it good enough to require `component(CardComponent) { "Hello world!" }` instead of `component CardComponent, "Hello world!"`? The block/Proc will be needed anyway, and then mixing initialize and template arguments is no longer a problem.