@joeldrapper 2022-08-12
> The called_from = caller_locations(1,1)[0].label.to_sym in Overrides#template is another slow caller info access.

We could get rid of that override by renaming the method to something that doesn't conflict with `<template>` tags, or having a special case method for `<template>` tags.

Back to the issue of this PR, we could just forget the implicit first-argument content feature or only allow keyword arguments from component initialisers. Either is acceptable IMO though not ideal.

I am concerned by this though.

> Re Proc#binding there are ideas floating around to remove it, because it causes unnecessary overhead for Ruby VMs.

We need a to-procable that yields to the first caller and the only I can see to do that is to use our own block type that holds a reference to the context (`self`) from the first use.