@eregon 2022-08-12
> We need a to-procable that yields to the first caller and the only I can see to do that is to use our own block type that holds a reference to the context (`self`) from the first use.

So you need the caller self but don't want to pass it explicitly, and just `self` is not what you want, right?
Potentially you could capture the caller self differently, e.g., with an explicit stack or save/restore pattern.
Maybe there should be `Proc#receiver` or so.

Why Proc#binding is problematic for optimization is:
```
def foo
  -> a, b { a + b }
end
```
Even such a lambda needs to capture the foo frame, because Proc#binding exposes it.
Without Proc#binding, and with a way to check at parse time for `Kernel#binding` and `*eval`, we wouldn't need to capture the foo frame here for the lambda. And even if the block captured variables we could just have a box to be able to write those captured variables but not need the full frame.