@joeldrapper 2022-08-12
Yes, the methods that accept blocks check if the block is already a `Phlex::Block` by checking the `binding.receiver`. If not a `Phlex::Block`, they create a `Phlex::Block` which stores the original block and `self`.

```ruby
if block_given? && !block.binding.receiver.is_a?(Phlex::Block)
  block = Phlex::Block.new(self, &block)
end
```

Phlex blocks respond to `to_proc` so can be coerced with `&` and passed down many times as an anonymous block. When the block is yielded, that hits `Phlex::Block#call` which executes the original captured block against the original captured context.

```ruby
def call(*args, **kwargs)
  @context.instance_exec(*args, **kwargs, &@block)
end
```

`Proc#receiver` would be sufficient for our use-case. Or perhaps a native way to “instance_yield” a block.